+++ src/routes/docs/components/Tooltip/+page.md
@@ -31,7 +31,7 @@ docUrl: $docUrl
 	import Labels from '$lib/components/Labels.svelte';
 	import Points from '$lib/components/Points.svelte';
 	import TooltipContext from '$lib/components/TooltipContext.svelte';
-	import TooltipContainer from '$lib/components/TooltipContainer.svelte';
+	import Tooltip from '$lib/components/Tooltip.svelte';
 	import TooltipItem from '$lib/components/TooltipItem.svelte';
 	import TooltipSeparator from '$lib/components/TooltipSeparator.svelte';
 
@@ -149,9 +149,9 @@ docUrl: $docUrl
 						<HighlightRect {...charts[0].axis && { axis: charts[0].axis}} />
 					{/if}
 				</Svg>
-				<TooltipContainer header={data => format(data.date, 'eee, MMMM do')} let:data>
+				<Tooltip header={data => format(data.date, 'eee, MMMM do')} let:data>
 					<TooltipItem label="value" value={formatNumberAsStyle(data.value, 'integer')} />
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
@@ -241,11 +241,11 @@ docUrl: $docUrl
 						<HighlightRect {...charts[1].axis && { axis: charts[1].axis}} />
 					{/if}
 				</Svg>
-				<TooltipContainer header={data => format(data.data.date, 'eee, MMMM do')} let:data>
+				<Tooltip header={data => format(data.data.date, 'eee, MMMM do')} let:data>
 					{#each keys as key}
 						<TooltipItem label="{key}" value={formatNumberAsStyle(data.data[key], 'integer')} />
 					{/each}
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
@@ -326,9 +326,9 @@ docUrl: $docUrl
 						<HighlightRect {...charts[2].axis && { axis: charts[2].axis}} />
 					{/if}
 				</Svg>
-				<TooltipContainer header={data => data.name} let:data>
+				<Tooltip header={data => data.name} let:data>
 					<TooltipItem label="date" value={format(data.startDate, 'h:mm a')} />
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
@@ -410,14 +410,14 @@ docUrl: $docUrl
 						<HighlightRect {...charts[3].axis && { axis: charts[3].axis}} />
 					{/if}
 				</Svg>
-				<TooltipContainer header={data => data.name} let:data>
+				<Tooltip header={data => data.name} let:data>
 					<TooltipItem label="start" value={format(data.startDate, 'h:mm a')} />
 					<TooltipItem label="end" value={format(data.endDate, 'h:mm a')} />
 					<TooltipSeparator />
 					<TooltipItem label="duration" valueAlign="right">
 						<Duration start={data.startDate} end={data.endDate} />
 					</TooltipItem>
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
@@ -499,14 +499,14 @@ docUrl: $docUrl
 						<HighlightRect {...charts[4].axis && { axis: charts[4].axis}} />
 					{/if}
 				</Svg>
-				<TooltipContainer header={data => data.name} let:data>>
+				<Tooltip header={data => data.name} let:data>
 					<TooltipItem label="start" value={format(data.startDate, 'h:mm a')} />
 					<TooltipItem label="end" value={format(data.endDate, 'h:mm a')} />
 					<TooltipSeparator />
 					<TooltipItem label="duration" valueAlign="right">
 						<Duration start={data.startDate} end={data.endDate} />
 					</TooltipItem>
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
@@ -588,9 +588,9 @@ docUrl: $docUrl
 						<HighlightRect {...charts[5].axis && { axis: charts[5].axis}} />
 					{/if}
 				</Svg>
-				<TooltipContainer header={data => format(data.date, 'eee, MMMM do')} let:data>
+				<Tooltip header={data => format(data.date, 'eee, MMMM do')} let:data>
 					<TooltipItem label="value" value={formatNumberAsStyle(data.value, 'integer')} />
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
     	</Chart>
     </div>
@@ -673,10 +673,10 @@ docUrl: $docUrl
 						<HighlightRect {...charts[6].axis && { axis: charts[6].axis}} />
 					{/if}
 				</Svg>
-				<TooltipContainer header={data => format(data.date, 'eee, MMMM do')} let:data>
+				<Tooltip header={data => format(data.date, 'eee, MMMM do')} let:data>
 					<TooltipItem label="value" value={formatNumberAsStyle(data.value, 'integer')} />
 					<TooltipItem label="baseline" value={formatNumberAsStyle(data.baseline, 'integer')} />
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
@@ -754,10 +754,10 @@ docUrl: $docUrl
 						<HighlightRect {...charts[7].axis && { axis: charts[7].axis}} />
 					{/if}
 				</Svg>
-				<TooltipContainer let:data>
+				<Tooltip let:data>
 					<TooltipItem label="x" value={formatNumberAsStyle(data.x, 'decimal')} />
 					<TooltipItem label="y" value={formatNumberAsStyle(data.y, 'decimal')} />
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
