+++ src/routes/docs/examples/Area/+page.md
@@ -18,7 +18,7 @@ docUrl: $docUrl
 	import HighlightLine from '$lib/components/HighlightLine.svelte';
 	import Labels from '$lib/components/Labels.svelte';
 	import TooltipContext from '$lib/components/TooltipContext.svelte';
-	import TooltipContainer from '$lib/components/TooltipContainer.svelte';
+	import Tooltip from '$lib/components/Tooltip.svelte';
 	import TooltipItem from '$lib/components/TooltipItem.svelte';
 
 	import Preview from '$lib/docs/Preview.svelte';
@@ -84,9 +84,9 @@ docUrl: $docUrl
 					<Area line={{ width: 2 }} />
 					<HighlightLine color="var(--color-blue-500)" />
 				</Svg>
-				<TooltipContainer header={data => format(data.date, 'eee, MMMM do')} let:data>
+				<Tooltip header={data => format(data.date, 'eee, MMMM do')} let:data>
 					<TooltipItem label="value" value={formatNumberAsStyle(data.value, 'integer')} />
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
@@ -140,9 +140,9 @@ docUrl: $docUrl
 					<Labels format="integer" />
 					<HighlightLine color={fruitColors[data?.fruit]} />
 				</Svg>
-				<TooltipContainer header={data => format(data.date, 'eee, MMMM do')}>
+				<Tooltip header={data => format(data.date, 'eee, MMMM do')}>
 					<TooltipItem label={data.fruit} value={formatNumberAsStyle(data.value, 'integer')} />
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
