+++ src/routes/docs/examples/AreaStack/+page.md
@@ -19,7 +19,7 @@ docUrl: $docUrl
 	import HighlightLine from '$lib/components/HighlightLine.svelte';
 	import Labels from '$lib/components/Labels.svelte';
 	import TooltipContext from '$lib/components/TooltipContext.svelte';
-	import TooltipContainer from '$lib/components/TooltipContainer.svelte';
+	import Tooltip from '$lib/components/Tooltip.svelte';
 	import TooltipItem from '$lib/components/TooltipItem.svelte';
 
 	import Preview from '$lib/docs/Preview.svelte';
@@ -92,11 +92,11 @@ docUrl: $docUrl
 					<AreaStack line={{ width: 2 }} />
 					<HighlightLine color="var(--color-blue-500)" />
 				</Svg>
-				<TooltipContainer header={data => format(data.data.date, 'eee, MMMM do')} let:data>
+				<Tooltip header={data => format(data.data.date, 'eee, MMMM do')} let:data>
 					{#each keys as key}
 						<TooltipItem label={key} value={formatNumberAsStyle(data.data[key], 'integer')} />
 					{/each}
-				</TooltipContainer>
+				</Tooltip>
 			</TooltipContext>
 		</Chart>
 	</div>
